+++ girder_annotation/girder_large_image_annotation/web_client/annotations/geojs/types/ellipse.js
@@ -1,0 +1,13 @@
+import rectangle from './rectangle';
+
+/**
+ * Convert a geojs ellipse annotation to the large_image
+ * annotation schema.
+ */
+function ellipse(annotation) {
+    const element = rectangle(annotation);
+    element.type = 'ellipse';
+    return element;
+}
+
+export default ellipse;