+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Features
+- Add support for ellipse and circle annotations
+
 ### Improvements
 - Improve parsing OME TIFF channel names ([806](../../pull/806))
 - Improve handling when a file vanishes ([807](../../pull/807))